tetra: move code from debug map command under pkg/bugtool
There should be no functional change, it's only so that I can reuse that
code in the bugtool command in the next patch.
